Refactor `coqLsp` tests according to review of PR#46
