+++ src/core/completionGenerator.ts
@@ -3,7 +3,10 @@ import { GeneratedProof } from "../llm/llmServices/generatedProof";
 
 import { CoqLspTimeoutError } from "../coqLsp/coqLspTypes";
 
-import { CompletionAbortError, throwOnAbort } from "../extension/extensionAbortUtils";
+import {
+    CompletionAbortError,
+    throwOnAbort,
+} from "../extension/extensionAbortUtils";
 import { EventLogger } from "../logging/eventLogger";
 import { asErrorOrRethrow, buildErrorCompleteLog } from "../utils/errorsUtils";
 import { stringifyAnyValue } from "../utils/printers";
