+++ src/test/resources/small_document.v
@@ -1,11 +1,11 @@
 Theorem test_thr : forall n:nat, 0 + n = n.
 Proof.
     intros n. Print plus.
-    auto.
-Admitted.
+    reflexivity.
+Qed.
 
 Lemma test_thr1 : forall n:nat, 0 + n + 0 = n.
 Proof.
     intros n. Print plus.
-    auto.
+    admit.
 Admitted.
